% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registryimporter-package.R
\docType{package}
\name{registryimporter-package}
\alias{registryimporter}
\alias{registryimporter-package}
\title{registryimporter: Registry Importer}
\description{
Create import files from study registries for Endnote databases.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/IQWiG/registryimporter}
  \item Report bugs at \url{https://github.com/IQWiG/registryimporter/issues}
}

}
\author{
\strong{Maintainer}: Claudia Kapp \email{claudia.kapp@iqwig.de}

}
\keyword{internal}
